pyproject.toml
[project]
name = "one-piece-lcd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "face-recognition>=1.3.0",
    "opencv-python>=4.8.0,<4.11",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "click>=8.1.0",
    "face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models",
    "setuptools>=80.9.0",
    # Pinned for mm* ecosystem compatibility (CPU mode for now)
    "torch==2.1.0",
    "torchvision==0.16.0",
    "transformers>=4.30.0",
    "numpy>=1.24.0,<2",
    "pillow>=10.0.0",
    # mm* packages for anime-face-detector (requires Python 3.11)
    "mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu121/torch2.1.0/mmcv_full-1.7.2-cp311-cp311-manylinux1_x86_64.whl ; python_version == '3.11'",
    "mmdet==2.28.0",
    "mmpose==0.29.0",
    "anime-face-detector>=0.0.9",
    "Cython>=3.0.0",
]

[project.scripts]
oplcd = "one_piece_lcd.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
# chumpy and xtcocotools need these during build
extra-build-dependencies = { chumpy = ["pip"], xtcocotools = ["numpy", "Cython"] }
preview = true
